+++ docs/.vitepress/config.js
@@ -16,6 +16,62 @@ export default defineConfig({
     config: MermaidPlugin,
   },
 
+  head: [
+    [
+      'link',
+      {
+        rel: 'apple-touch-icon',
+        sizes: '180x180',
+        href: 'apple-touch-icon.png',
+      },
+    ],
+    [
+      'link',
+      {
+        rel: 'icon',
+
+        href: 'favicon.ico',
+      },
+    ],
+    [
+      'link',
+      {
+        rel: 'icon',
+        type: 'image/png',
+        sizes: '32x32',
+        href: 'favicon-32x32.png',
+      },
+    ],
+    [
+      'link',
+      {
+        rel: 'icon',
+        type: 'image/png',
+        sizes: '16x16',
+        href: 'favicon-16x16.png',
+      },
+    ],
+    ['link', { rel: 'manifest', href: 'site.webmanifest' }],
+    [
+      'link',
+      {
+        rel: 'mask-icon',
+        href: 'safari-pinned-tab.svg',
+        color: '#3a0839',
+      },
+    ],
+    ['link', { rel: 'shortcut icon', href: 'favicon.ico' }],
+    ['meta', { name: 'msapplication-TileColor', content: '#3a0839' }],
+    [
+      'meta',
+      {
+        name: 'msapplication-config',
+        content: 'browserconfig.xml',
+      },
+    ],
+    ['meta', { name: 'theme-color', content: '#ffffff' }],
+  ],
+
   themeConfig: {
     // logo: "https://fairdataihub.org/logo.svg",
     // siteTitle: false,
@@ -118,6 +174,7 @@ function sidebarGuide() {
       collapsible: true,
       items: [
         { text: 'Commenting code', link: '/general/comments.md' },
+        { text: 'Project setup', link: '/general/project-setup.md' },
         { text: '@fairdataihub/config', link: '/general/config.md' },
       ],
     },